requirements.txt
django
djangorestframework
psycopg2-binary 
socket
os
dotenv